fix: change data type of editHistory.editType column

+++ api/src/model/migration/1591519630478-edit_histories.ts
@@ -10,7 +10,7 @@ export class editHistories1591519630478 implements MigrationInterface {
                 \`user_id\`    bigint             NULL,
                 \`ref_id\`     bigint             NULL,
                 \`ref_type\`   bigint             NULL,
-                \`edit_type\`  bigint             NULL,
+                \`edit_type\`  smallint             NULL,
                 \`created_at\` DATETIME           NOT NULL default now(),
                 \`updated_at\` DATETIME           NOT NULL default now(),
                 \`deleted_at\` DATETIME           NULL
